Revert "envsetup.sh: explain _why_ there's an adb function in here."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b0fd8e09-6485-4835-9fd1-0386f9338a5c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica7a6ca12bb9df257739c6342d1a9f22b91b01d0