Add upload_log to function adb
The purpose of the upload_log function is to collect basic
information about the adb run (e.g. the command passed, username
etc.) and upload the info to Sawmill for further analysis.
More details can be found in go/local-dev-tools-instrumentation

The function is a no-op in aosp, the actual implementation will
live under vender/google/tools and is only visible for intenal
users.

Test: manual
Change-Id: I001ba3c6c30b3ffc95d0fdb30ea7178a991c680f