Revert "Migrate buildinfo.sh script into Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2893b550-8816-433d-b43d-dbc9b10124ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I886596beef55f833be1acac8a87d8f023e4860df