Revert "Copy build.prop[-<part>] to dist dir only if the corresp..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e05a9ddf-3bba-4e68-94b9-63ee161bf217). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieae979784ede0c1a7cf86defe14152ba43ea1ff6