Revert "16k: core: Enable TARGET_NO_BIONIC_PAGE_SIZE_MACRO by de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69205dce-a1d9-4487-bc7c-1779e655052a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I89b5b8cff70d073efe094694fb2ed1827d4c17de