Revert "16k: core: Enable TARGET_NO_BIONIC_PAGE_SIZE_MACRO by de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/424731e1-e841-4d91-94fc-89203c1fc6a6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic081ab9f4ec4e447bf6410d8c80ad126c0ed4332