Revert "Copy build.prop[-<part>] to dist dir only if the corresp..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69b9aa60-7b29-43f6-93cc-d789661ba111). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb024370112d9daa6aa3dad89db9f1960f35f439