Revert "16k: core: Enable TARGET_NO_BIONIC_PAGE_SIZE_MACRO by de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ae01630-bbb2-4700-93b9-fa722e3910fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5e47fd965ad7b9d1cfd4fe251b95a418464c756f