Revert "16k: core: Enable TARGET_NO_BIONIC_PAGE_SIZE_MACRO by de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/305e43b4-1817-4a97-ad75-9e2cf7218405). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I70ac213c827bf0e94613b4696672a11cd8543e62