Revert "16k: core: Enable TARGET_NO_BIONIC_PAGE_SIZE_MACRO by de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/44625c3d-3612-47c1-b191-778d934147b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f076b628c0246b2e0f796408dd0d6c693188729