aconfig: temporarily to turn off cotnainer filed assignment to bypass performance regression.
Context: we detected performance regression on some devices when
setting container fields on aconfig packages. This regression is in the
way of turning on container field enforcement on both
aconfig_declarations and .aconfig definitions. One way to bypass this
regression is to not include the container field in the generated
aconfig caches. This has been verified on ABTD. We will revret this cl
once we have figured out the underlying root cause of the performance
regression.

Bug: b/333868241
Test: atest aconfig.test
Change-Id: Ia04e9f8cce6cd6e1b0f79c837be4fe5b52f9dd4e