Revert "Expose RELEASE_AVF_SUPPORT_CUSTOM_VM_WITH_PARAVIRTUALIZE..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ac9376a-9da5-4991-b060-72d772a9154f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa1ccc654ebc916f479d7252b615fec726d8560c