Revert "Expose RELEASE_AVF_SUPPORT_CUSTOM_VM_WITH_PARAVIRTUALIZE..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b65cb5f6-8490-49b2-901c-ce462e9e5de1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9e6a442ffb36dcdd64edf9c3d556ab0c9f7ce10