Add dependencies for module which define LOCAL_SOONG_INSTALL_PAIRS
Soong already generated phony target for module which defined
LOCAL_SOONG_INSTALL_PAIRS but the dependencies is not expose to make
which lead to the problem that only the last one in
LOCAL_SOONG_INSTALL_PAIRS be installed if m that target. This CL force
add the phony target as dependencies which is generated by soong and kati.

Bug: 333974089
Test: apply aosp/3041672 ; m hyph-lic
Change-Id: I79050e99c3e63f58cfd32db270c909a76bd96801