Revert "Define a make target for all APEX dumps"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7eb33a7b-2d8a-4107-9c24-5dbb9b3291f7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0c9f80c4e66dde775c91b3559c2b29270affc93