Revert "Define a make target for all APEX dumps"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9fd154d6-e068-4d12-b286-b4c3f9fd9378). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38f69cbfbb7d6d0c85e58df1c231973121d76d7e