Revert "Revert^2 "Expose RELEASE_AVF_SUPPORT_CUSTOM_VM_WITH_PARA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb21f2d9-5a8f-4d75-b72f-5e6f3f5cc8fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I526e0d90d2d6bc8cc57f4af82db5e3b8854b3701