Revert "16k: build: Enable TARGET_NO_BIONIC_PAGE_SIZE_MACRO by d..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d8ce27e-41ef-48ec-a474-e8ebe245757d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I874647871cff1337c2c77a02423032210651ba7e