Revert "Update aconfig storage deps for CPP codegen"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f25c28ec-9b47-49d8-9642-b5fb9743cbd4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I812b1b67bcd1a17f517dcbcef3304155b7127aef