Revert "Update aconfig storage deps for CPP codegen"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed2dfa55-001c-4a21-9300-102dd1995c70). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I82abe420eb7f728ebf5c9683d72b9f1faeb40af5