Revert "aconfig: update flag info storage file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2805a5c7-974a-404f-a928-37eab3fa91ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8dd44655300f50c3f7fda1d6bc9e8b951c4ef752