Revert "Update aconfig storage deps for CPP codegen"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f958ccb1-60cd-4e61-8f6e-2dfa3868d926). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id74c1a4785deecc81456fc7296f779f2f372dd04