Revert "Drop cmdgoals special case from prebuilt selection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d4fda06e-1da1-481f-87b2-606062ec9435). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic2c967bdc983d25b700de5d5d5f830702bed62b5