Revert "Drop cmdgoals special case from prebuilt selection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca30fbc4-1741-4a3e-863d-7f41a59fd455). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I742ea422901e2b45559bef5ba94706544607fb12