Revert "Drop cmdgoals special case from prebuilt selection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/333c4e9c-e857-4912-9ecf-533edcbdda28). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idaf1e4c140d37fa4dcf0fe2c7b2a337c72074906