Revert "Use a version script for libaconfig_storage_read_api_cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/857f0d40-a34b-46c0-8d2b-3e6a76df4a8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia865fbe0e656881b85377ecdccebc96ea877a99d