Revert "Generate CustomFeatureFlags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cccc257a-8524-4176-ba7b-e04adc8dfc0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id39e33d0aae8a766062564987f49dd133a88c7c1