Revert "Update aconfig storage deps for CPP codegen"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a57486e-61a5-466f-9071-19a9e05ea61a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8700eb8202cfa647b4b0c6b3d4aefce4a416506f