Revert "Update aconfig storage deps for CPP codegen"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/798b8443-12ad-481f-acc5-dc10a2effdc3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I248a37f77d178d35ed409d1fb5e015c99c513fc4