Revert "Update aconfig storage deps for CPP codegen"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/09094539-2817-43e6-b932-31d0a2810196). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I76f88bee082820a3b4e805910a85eede031e6377