Revert "Align boot jar exclusion with apex_contributions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34c64296-0e72-457a-b4d2-b700c6c13e06). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If60c8f711de3810de53572b6a40b8a08d8fffae3