check-flagged-apis: ignore interfaces when parsing API signature files
Each symbol in the output from parsing API signature files is verified
to exist in the api-versions.xml if the flag is enabled, and vice versa.
However, interfaces are never instantiated so it is an error to include
them in the output of parseApiSignature. Explicitly skip interfaces when
encountered during parsing.

Bug: 334870672
Test: croot && build/tools/check-flagged-apis/check-flagged-apis.sh
Change-Id: Id4bd3af4174bf1b36f36401064637bd49ae9c176