Revert "Remove libmdnssd"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/52acbd3c-0fc1-4699-ae5b-7724550cbb0f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 341118115
Change-Id: I04fda15b06685a4ccf00a443e82bdef897d61cb3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com