Revert "Move all GKI targets away from non-AB"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/093cdcc5-c126-4b40-8255-1fb115322bbe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 341139770
Change-Id: Idc7c65d6cdd27712be3074eb4e95fe448adbc1a9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com