Revert "Add PRODUCT_16K_DEVELOPER_OPTION to product.mk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/978ba457-60c6-4e83-833f-324674933cb3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b5d0344fb4111265bf3819c09f518ee48dcf5c3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com