aconfig: Make cpp code for test compatible with prod
The usecase is too be able to compile everything with the
prod cpp codegen and only swap in the test library for the test
by using the fact that dependencies in soong are not transitive.
meaning making work:
  libA depends on flag_cpp_lib
  test_for_libA depends on libA and flag_cpp_test_lib

Currently this does not work because the definition of
flag_provider_interface is different between the test
and prod codegen and so their abi (vtable) is not compatible.

To make this possible this CL put the extra test method into
another child class so that the parent class
(flag_provider_interface) has the same definition in test
and prod.

Test: cargo test
Test: atest net_test_bta with https://r.android.com/c/3094400/1
Bug: 341965901
Change-Id: I2e472241c9b5f023921a2c45ed44187321ce4d9c